refactor(editor): move useDebounce and durations into @n8n/composables
git mv useDebounce (+test) and durations (+test) from editor-ui's app
kernel into @n8n/composables; useDebounce's durations import is repointed
to the co-located module. Kernel slice v1 (CAT-3686). Shims follow.
