+++ packages/frontend/@n8n/composables/src/durations.ts
